Create mode-plantuml.js
Hello,

I don't know if this would be the best way to contribute with a new syntax file... But I didn't find an alternative. :)

Here's a first version (with minor little issues, and with some tokens still missing). It works... :D I've tested on SlimText and it works as expected (I'm still refining it) and I'd like you to take a look and tell me if it's ok.

http://www.plantuml.com  to get some sample codes to test.